[SPARK-54023][CORE] Support `AUTO` IO Mode
### What changes were proposed in this pull request?

This PR aims to support a new Netty IO Mode, `AUTO`, on top of the existing `NIO`, `EPOLL`, and `KQUEUE`.

`AUTO` mode prefers to use native `EPOLL` mode on Linux and `KQUEUE` mode on MacOS if available. Then, it fallbacks to `NIO` mode.

### Why are the changes needed?

To help a user to try to use native IO mode more easily.

### Does this PR introduce _any_ user-facing change?

No, this is a new IO mode.

### How was this patch tested?

Pass the CIs with newly added test suite, `ShuffleNettyAutoSuite`.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #52724 from dongjoon-hyun/SPARK-54023.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>