[SPARK-54040][BUILD] Remove unused commons-collections 3.x
### What changes were proposed in this pull request?

The latest commons-collections 3.x was published in Nov 2015, which is pretty old, it was added back to Spark dep list in SPARK-38696, due to Hadoop transitive usage.

HADOOP-15760 (fixed in 3.4.2, 3.5.0) totally removes usage of it, so it's safe to remove it from Spark now.

### Why are the changes needed?

Remove useud dependency.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Pass GHA.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #52743 from pan3793/SPARK-54040.

Authored-by: Cheng Pan <chengpan@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>