[SPARK-53732][SQL] Remember TimeTravelSpec in DataSourceV2Relation
### What changes were proposed in this pull request?

This PR adds `TimeTravelSpec` to `DataSourceV2Relation` when the relation is created by time traveling.

### Why are the changes needed?

These changes are needed for subsequent PRs where I will modify Spark to reload certain tables to ensure consistent version scanning and DELETE, UPDATE, and MERGE isolation. Without this change, Spark looses track of whether the relation points to the current version of the table or time travels. As an engine, Spark must be aware whether a relation is the result of time traveling or points to the current table version/snapshot.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

This PR comes with tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #52599 from aokolnychyi/spark-53732-v2.

Authored-by: Anton Okolnychyi <aokolnychyi@apache.org>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>