[SPARK-53891][SQL][FOLLOW-UP] Clarify javadocs for case when merge summary metric is not found
### What changes were proposed in this pull request?
Clarify javadocs to explain situations where the metric is not found.

### Why are the changes needed?
As we begin to handle more write summaries like in https://github.com/apache/spark/pull/52669/, involving more complex walks of the executed plan graph, the code to calculate merge summary may encounter some plan it does not expect and would need to populate -1.

This was actually called out in https://github.com/apache/spark/pull/52595#discussion_r2427651730 , it was not done as it was not case then, but now I realize it will be possible as this code evolves.  Especially as we plan to still populate MergeSummary in cases where the optimizer rewrites Merge plan to get rid of MergeRowsExec or Join.

As it is more an error-handling case, we don't need to change the model of the MergeSummary to return Long or OptionalLong, so we can put -1 and indicate this in the javadoc.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
No

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #52797 from szehon-ho/SPARK-53891-follow.

Authored-by: Szehon Ho <szehon.apache@gmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>