[SPARK-54106][SS] State store row checksum implementation
### What changes were proposed in this pull request?

This introduces row checksum creation and verification for state store, both HDFS and RocksDB state store. This will help detect corruption at the row level and help prevent us from corrupting the remote checkpoint. We also verify the checksum when loading from checkpoint to detect if a row in the checkpoint is corrupt.

Since this adds overhead, **it is disabled by default and can only be enabled for a new checkpoint** since the conf is also written to the offset log to ensure that.

This also introduces a readVerificationRatio conf, that allows checksum verification when a row is read by client. This makes corruption detection more immediate and will fail the query task, rather than when it is too late. But can be expensive to do this frequently for reads, hence why the frequency is configurable.

### Why are the changes needed?

Integrity verification

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

New tests added

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #52809 from micheal-o/row_checksum.

Authored-by: micheal-o <micheal.okutubo@gmail.com>
Signed-off-by: Anish Shrigondekar <anish.shrigondekar@databricks.com>