[SPARK-54121][SS] Automatic Snapshot Repair for State store
### What changes were proposed in this pull request?

Today, the engine currently treats both the changelog and snapshot files with the same importance, so when a state store needs to be loaded it reads the latest snapshot and applies the subsequent changes on it. If the snapshot is bad or corrupted, then the query will fail and be completely down and blocked, needing manual intervention. This leads to user clearing their query checkpoint and having to do full recomputation.

This shouldn’t be the case. The changelog should be treated as the “source of truth” and the snapshot is just a disposable materialization of the log.

Introducing Automatic snapshot repair, which will automatically repair the checkpoint by skipping bad snapshots and rebuilding the current state from the last good snapshot (works even if there’s none) and applying the changelogs on it. This eliminates the need for manual intervention and unblocks the pipeline to keep it running.

Also emit metrics about number of state stores that were auto repaired in a given batch, so that you can build alert and dashboard for it.

### Why are the changes needed?

Automatic failure recovery

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Added new tests

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #52821 from micheal-o/auto_snapshot_repair.

Authored-by: micheal-o <micheal.okutubo@gmail.com>
Signed-off-by: Anish Shrigondekar <anish.shrigondekar@databricks.com>