[SPARK-54184][K8S] Support `spark.kubernetes.executor.deletedExecutorsCacheTimeout`
### What changes were proposed in this pull request?

This PR aims to support `spark.kubernetes.executor.deletedExecutorsCacheTimeout`.

### Why are the changes needed?

To allow users to control the TTL for the deleted executors cache.

Previously, it has a hard-coded value. In some very slow clusters, we need to remember longer than 3 minutes.

https://github.com/apache/spark/blob/a8e35c407bc5340f83b35e5a2f0b0767c6baadb0/resource-managers/kubernetes/core/src/main/scala/org/apache/spark/scheduler/cluster/k8s/ExecutorPodsLifecycleManager.scala#L54-L57

### Does this PR introduce _any_ user-facing change?

No behavior change because the default value is the same.

### How was this patch tested?

Pass the CIs.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #52884 from dongjoon-hyun/SPARK-54184.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>