[SPARK-54178][SQL] Improve error for ResolveSQLOnFile
### What changes were proposed in this pull request?

Improve error for ResolveSQLOnFile - generic error does not mean that the data source is not supported!

### Why are the changes needed?

Currently `ResolveSQLOnFile` throws `UNSUPPORTED_DATASOURCE_FOR_DIRECT_QUERY` for a generic failure when discovering files and figuring out the file schemas. This is confusing. We need a separate error.

### Does this PR introduce _any_ user-facing change?

Better error message.

### How was this patch tested?

Hard to create a test, because the files need to be corrupted.

### Was this patch authored or co-authored using generative AI tooling?

Claude.

Closes #52875 from vladimirg-db/vladimir-golubev_data/better-error-for-resolve-sql-on-file.

Authored-by: Vladimir Golubev <vladimir.golubev@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>