[SPARK-54217] Synchronize PythonRunner's MonitorThread kill decision
This diff addresses the synchronization issue described in SPARK-54217 by respecting the existing
releasedOrClosed AtomicBoolean in the PythonRunner's kill codepath, which is currently only used in
the "released" codepath - not the "closed" one. In doing so, we avoid erroneously destroying a
still-healthy Python worker; in the current state, it will be destroyed & a new one will be created.