[SPARK-54200][SS] Call close() against underlying InputPartition when LowLatencyReaderWrap.close() is called
### What changes were proposed in this pull request?

This PR proposes to fix the bug of missing close() on underlying InputPartition when LowLatencyReaderWrap.close() is called.

### Why are the changes needed?

Not closing the underlying InputPartition could leak resource; e.g. Kafka consumer is not returned to the pool, which ends up with destroying the purpose of connection pool and creating Kafka consumer instances every batch.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

A new UT for Kafka rather than general one, since Kafka data source has an internal metric to provide the necessary information for validation.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #52903 from HeartSaVioR/SPARK-54200.

Authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>