[SPARK-54242][BUILD] Skip `Checkstyle` if `NOLINT_ON_COMPILE` is true
### What changes were proposed in this pull request?

This PR aims to skip `Checkstyle` if `NOLINT_ON_COMPILE` is true.

### Why are the changes needed?

Like `Scalastyle`, `Checkstyle` is also a kind of linting. So, we had better skip it when a user set `NOLINT_ON_COMPILE` to `true`.

### Does this PR introduce _any_ user-facing change?

No Spark behavior change because this is only a build change.

### How was this patch tested?

Pass the CIs and manually.

```
$ NOLINT_ON_COMPILE=true build/sbt package | grep checkstyle
```

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #52943 from dongjoon-hyun/SPARK-54242.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>