[SPARK-53823][SS] Implement allow list for real time mode
### What changes were proposed in this pull request?

Add an allowlist and some guardrails to help users understand what is supported and what is not supported in Real-time Mode.  This should improve the user experience of real-time mode and minimize the chance of a new user use it in a way that is unexpected or untested which may produce undesirable outcomes.

### Why are the changes needed?

To help guide users on what is currently supported and what is not.

### Does this PR introduce _any_ user-facing change?

Yes, running a query with currently unsupported features in RTM will throw an exception.

### How was this patch tested?

Several tests are added

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #52891 from jerrypeng/SPARK-53823.

Authored-by: Jerry Peng <jerry.peng@databricks.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit 178f0f428020dd1795313afe1c6dd0c125e1b70d)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>