[MINOR][INFRA] Fix the install version comment to Python 3.10 in python-minimum Dockerfile
### What changes were proposed in this pull request?

This PR aims to fix a comment with `Python 3.10` instead of `Python 3.9` according to the code inside `python-minimum/Dockerfile`.

### Why are the changes needed?

This seems to be a left-over which we missed.

### Does this PR introduce _any_ user-facing change?

No. This is a comment change in the infra Dockerfile.

### How was this patch tested?

Manual review.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #52946 from dongjoon-hyun/minor-minimum.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit 4ec6668a688cac095773f6623ac5509d3b1552c5)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>