[SPARK-54239][BUILD] Bump Guava 33.4.8
### What changes were proposed in this pull request?

This PR bumps Guava from 33.4.0 to 33.4.8, the latest patch version of 33.4 serial.

The release notes can be found at

- https://github.com/google/guava/releases/tag/v33.4.8
- https://github.com/google/guava/releases/tag/v33.4.1 (this contains more info)

As mentioned in the release notes, it migrates from jsr305 to jspecify, thus causes some transitive deps changes

https://mvnrepository.com/artifact/com.google.guava/guava/33.4.0-jre
https://mvnrepository.com/artifact/com.google.guava/guava/33.4.8-jre

```patch
- com.google.code.findbugs:jsr305:3.0.2
- org.checkerframework:checker-qual:3.34.0
+ org.jspecify:jspecify:1.0.0
```

Note, the transitive deps change does not affect the Maven package because we use the white list to declare the included deps for shaded jars.

### Why are the changes needed?

The current Guava 33.4.0 used by Spark was released in Dec 17, 2024. This upgrades it to the latest patched version (instead of 33.5.0) to make it align with gRPC 1.76.0 (this is not enforced, but nice to have)

https://mvnrepository.com/artifact/io.grpc/grpc-api/1.76.0

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Pass GHA.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #52939 from pan3793/guava-33.4.8.

Authored-by: Cheng Pan <chengpan@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>