[SPARK-54263][INFRA] Fix `run-in-container` script to use Python 3.11 instead of 3.9
### What changes were proposed in this pull request?

This PR aims to fix `run-in-container` script to use Python 3.11 instead of 3.9.

### Why are the changes needed?

Since Apache Spark 4.0.0, we have an optional way to build docs in Docker image.
- #48860

https://github.com/apache/spark/blob/fa33ea000a0bda9e5a3fa1af98e8e85b8cc5e4d4/docs/README.md?plain=1#L129-L136

When we started to use Python 3.11 instead of 3.9 for Apache Spark 4.1.0, we missed to update this `run-in-container`.
- #51150
- #51198

### Does this PR introduce _any_ user-facing change?

No behavior change because this is an optional way to generate Spark docs.

### How was this patch tested?

Manually test.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #52959 from dongjoon-hyun/SPARK-54263.

Authored-by: Dongjoon Hyun <dongjoon@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit 6466ed579d5aeca31c94d0a0dbbad00865caa2c3)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>