[SPARK-54290][SS] Skip checksum creation if path already exists without checksum
### What changes were proposed in this pull request?

This PR modifies ChecksumCheckpointFileManager to fall back to the underlying CheckpointFileManager when the target path exists but the checksum file does not. This prevents failures in checkpoint recovery scenarios where checksum validation cannot be performed.

### Why are the changes needed?

Consider the case using STATE_STORE_CHECKPOINT_FORMAT_VERSION = 1 when a batch fails but state files are written. If on the next run, we try to upload both a new state file and a file checksum, the file could fail to be uploaded but the file checksum is uploaded successfully. This would lead to a situation where the old file could be loaded and compared with the new file checksum, which would fail the checksum verification. This issue does not happen when STATE_STORE_CHECKPOINT_FORMAT_VERSION = 2 since each batch run unique ids will be created.

### Does this PR introduce _any_ user-facing change?

### How was this patch tested?

- Added unit tests in ChecksumCheckpointFileManagerSuite to verify fallback behavior
- Added a failure injection test in RocksDBCheckpointFailureInjectionSuite to simulate how this error is caused

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #52985 from dylanwong250/SPARK-54290.

Authored-by: Dylan Wong <dylan.wong@databricks.com>
Signed-off-by: Anish Shrigondekar <anish.shrigondekar@databricks.com>
(cherry picked from commit 811482a79cbc9b26c1502cbaed114340e9a82379)