[SPARK-53917][CONNECT] Support large local relations - follow-ups
### What changes were proposed in this pull request?

This PR adds several fixes and improvements over https://github.com/apache/spark/pull/52613 which added support for 2GB+ local relations in Spark Connect.

#### Uploading batches of chunks
Currently, before caching the local relation on the server via `ChunkedCachedLocalRelation`, the client materializes all chunks in memory (1 schema chunk and N data chunks). This can lead to high memory pressure on the client when uploading very large local relations.

In this PR, I'm changing how the client uploads the local relation. Instead of materializing all chunks in memory, the client will materialize a batch of chunks in memory, upload the batch of chunks to the server, and proceed to collecting the next batch of chunks. The size of the batch of chunks is controlled via `spark.sql.session.localRelationBatchOfChunksSizeBytes` (1GB by default). This way, the uploading mechanism only consumes 1GB of memory at each point in time. Alternatives to this approach are:
a) uploading each chunk separately - would require one pair of `ArtifactStatuses` and `AddArtifactsRequest` RPC calls for each chunk, which is inefficient.
b) (current implementation) materializing all chunks in memory and uploading them via a single pair of `ArtifactStatuses` and `AddArtifactsRequest` RPC calls. This can lead to high memory pressure on the client.
Uploading batches of chunks is a middle-ground solution.

Changes are implemented both for the python and scala clients.

#### Minor fixes and improvements
- Replace `ArraySeq.unsafeWrapArray(data.map(_.copy()).toArray))` with `data.map(_.copy()).toArray.toImmutableArraySeq` in `SparkConnectPlanner.scala`. The latter is compatible with both scala 2.13 and scala 2.12 and is consistent with how arrays are converted to sequences in other places in the code base.
- Improved asserts and tests in the python client.

### Why are the changes needed?

Reduce memory pressure in the spark connect python and scala clients when uploading very large local relations to the server.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Existing tests.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #52973 from khakhlyuk/largelocalrelations-followup.

Authored-by: Alex Khakhlyuk <alex.khakhlyuk@gmail.com>
Signed-off-by: Herman van Hovell <herman@databricks.com>