[SPARK-54306] Annotate Variant columns with Variant logical type annotation
### What changes were proposed in this pull request?

This PR makes changes to the parquet writer to make it annotate variant columns with the parquet variant logical type annotation.

### Why are the changes needed?

The Parquet spec has formally adopted the Variant logical type, and therefore, Variant columns must be properly annotated in Spark 4.1.0 which depends on Parquet-java 1.16.0 which contains the variant logical type annotation.

This change is hidden behind a flag that is disabled by default until read support can be properly implemented.

### Does this PR introduce _any_ user-facing change?

Yes, Parquet files written by Spark 4.1.0 with the flag enabled (which it eventually will be by default) could contain the variant logical type annotation which readers without support for the type will not be able to read

### How was this patch tested?

Unit test to check if nested as well as top-level variants are properly annotated, and the data is being written correctly.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #53005 from harshmotw-db/harshmotw-db/variant_annotation_write.

Authored-by: Harsh Motwani <harsh.motwani@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>