[SPARK-54157][SQL] Fix refresh of DSv2 tables in Dataset
### What changes were proposed in this pull request?

This PR fixes refresh of DSv2 tables in Dataset.

### Why are the changes needed?

Prior to this change, Spark would pin the version of DSv2 tables at load/resolution time. Any changes within the session will not be propagated to the analyzed but not yet executed Dataset, breaking the behavior compared to DSv1 tables. Changes in this PR are needed for the following reasons:
- Prevent scanning/joining inconsistent versions of the table in the same session (see tests).
- Prevent stale results upon external and session changes.
- Remove workarounds in DSv2 connectors by fixing the problem in Spark.

### Does this PR introduce _any_ user-facing change?

Yes, but this PR makes DSv2 Table behavior match the expected Spark semantics.

### How was this patch tested?

This PR comes with tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #52920 from aokolnychyi/spark-54157.

Authored-by: Anton Okolnychyi <aokolnychyi@apache.org>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>