[SPARK-54358][SDP] Checkpoint dirs collide when streaming tables in different schemas have same name
### What changes were proposed in this pull request?

Updates the per-streaming table checkpoint path to use the fully qualified table path, instead of just its name.

### Why are the changes needed?

A streaming table is a table fed by a stream. Streaming tables have checkpoint directories underneath their pipeline's storage root. These directories don't currently take the table namespace into account, which means that two tables with different schemas but the same name will be mapped to the same checkpoint directory. This could be very bad and cause data loss.

### Does this PR introduce _any_ user-facing change?

Yes, but for an unreleased feature.

### How was this patch tested?

Added a test for the collision case. Verified that it fails with the prior logic and now passes.

### Was this patch authored or co-authored using generative AI tooling?

Closes #53070 from sryza/collide.

Authored-by: Sandy Ryza <sandy.ryza@databricks.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>