[SPARK-53924] Reload DSv2 tables in views created using plans on each view access
### What changes were proposed in this pull request?

This PR makes Spark reload DSv2 tables in views created using plans on each view access.

### Why are the changes needed?

The current problem is that the view definition in the session catalog captures the analyzed plan that references `Table` (that is supposed to pin the version). If a connector doesn’t have an internal cache and produces a new `Table` object on each load, the table referenced in the view will become orphan and there will be no way to refresh it unless that `Table` instance auto refreshes on each scan (super dangerous).

### Does this PR introduce _any_ user-facing change?

Yes, but it restores the correct behavior without requiring hacks in connectors.

### How was this patch tested?

This PR comes with tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #52876 from aokolnychyi/spark-53924.

Authored-by: Anton Okolnychyi <aokolnychyi@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>