[SPARK-54063][SS] Trigger snapshot for next batch when upload lag
## What changes were proposed in this pull request?
We are adding a functionality for StateStoreProvider to force creating a snapshot on commit when creating a StateStore when it detects that the corresponding StateStoreProvider is lagging in creating snapshot (i.e. having too many changelog). This applies to both RocksDB and HDFS. This feature is only enabled when `STATE_STORE_COORDINATOR_REPORT_SNAPSHOT_UPLOAD_LAG` is true as well (introduced in [PR 50123](https://github.com/apache/spark/pull/50123/files#diff-7c577967a171f51523afddd7a5eca49806432fae2da8dc35c114d9699e6e3e40))

We also added a StateStoreCustomMetric `CUSTOM_METRIC_FORCE_SNAPSHOT` to track the number of forced snapshot

The changes made in this PR includes
-  `StateStoreCoordinator` now keeps track of a list of lagging state stores and update it accordingly on receiveing `LogLaggingStateStores` (update the set) and `DeactivateInstances` (remove all lagging stores from set)
- Change `ReportActiveInstance` in StateStoreCoordinator to also return `shouldForceSnapshot`, which indicates if current `StateStore` should create a snapshot on next commit
- Add a new param `forceSnapshotOnCommit: Boolean  =  false` to StateStoreProvider function `getStore`, and its inherited provider can create state stores that will force a snapshot on commit
- When StateStore singleton `get()` is called, it will call `ReportActiveInstance` to check if it should create a snapshot in the next commit.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Added integration test in StateStoreCoordinatorSuite, where it tests if RocksDBStateStore and HDFSBackedStateStore can both auto-trigger a snapshot on commit when delay in snapshot upload is detected

Also added unit test in StateStoreBaseSuite to test if the metric is populated correctly when a snapshot is forced.

### Was this patch authored or co-authored using generative AI tooling?

Yes

Closes #52773 from zifeif2/trigger-snapshot.

Authored-by: Ubuntu <zifei.feng@your.hostname.com>
Signed-off-by: Anish Shrigondekar <anish.shrigondekar@databricks.com>