[SPARK-54393][BUILD][TESTS] Upgrade oracle free to 23.26.0
### What changes were proposed in this pull request?
This PR aims to upgrade the docker image for Oracle Database to `23.26.0`.

### Why are the changes needed?
Ensuring the JDBC source works with the newer Oracle Database.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Tested on my dev environment.
With `PASSWORD_INIT_TIMEOUT=30`, I observed the test failed 3 times out of 50 rounds due to timeout. So I increased the timeout to `60` and the test failed 0 times out of 50 rounds.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #53107 from sarutak/upgrade-oracle-free-23.26.

Authored-by: Kousuke Saruta <sarutak@amazon.co.jp>
Signed-off-by: yangjie01 <yangjie01@baidu.com>