[SPARK-54349][PYTHON] Refactor code a bit to simplify faulthandler integration extension
### What changes were proposed in this pull request?
There are many places where pyspark is trying to integrate with faulthandler and use the same functionality to dump stack traces/record thread dumps.
In order to reduce the complexity of the integration and ease the extension of integration it makes sense to technically refactor the code to use the same code in all the places.

### Why are the changes needed?
Improves developer experience.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

By existing unit tests

### Was this patch authored or co-authored using generative AI tooling?

Closes #53016 from antban/simplify-faulthandler-integration.

Authored-by: antban <dmitry.sorokin@gmail.com>
Signed-off-by: Takuya Ueshin <ueshin@databricks.com>