[SPARK-54346][SS] Introduce state repartition API and repartition runner
### What changes were proposed in this pull request?

Introducing the API for offline repartitioning of streaming state. This is currently not exposed, since it is still in development. Also implemented some of the core functionalities of the repartition batch runner, that validates the checkpoint, creates the repartition batch and commits. Subsequent PRs will build on this. Also Spark connect and pyspark APIs will be added in subsequent PRs.

Also introduce the streamingCheckpointManager for performing operations on the streaming checkpoint. This is currently not exposed, since it is still in development.

### Why are the changes needed?

Streaming state repartitioning

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

New test suite added

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53056 from micheal-o/repartition_api.

Authored-by: micheal-o <micheal.okutubo@gmail.com>
Signed-off-by: Anish Shrigondekar <anish.shrigondekar@databricks.com>