[SPARK-53322][SQL][4.1] Select a KeyGroupedShuffleSpec only when join key positions can be fully pushed down
### What changes were proposed in this pull request?
When a KeyGroupedShuffleSpec is used to shuffle another child of a JOIN, we must be able to push down JOIN keys or partition values to be able to ensure that both children have matching partitioning. If one child reports a KeyGroupedPartitioning but we can't push down these values (for example, if the child was a key-grouped scan that was checkpointed), then this information cannot be pushed down to the child scan and we should avoid using this shuffle spec to shuffle other children.

### Why are the changes needed?
Prevents potential correctness issue when key-grouped partitioning is used on a checkpointed RDD.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
See test changes.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #53132 from chirag-s-db/checkpoint-pushdown-4.1.

Authored-by: Chirag Singh <chirag.singh@databricks.com>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>