[SPARK-54416][DOCS] Remove references to Pod Security Policies from Kubernetes docs
### What changes were proposed in this pull request?
The Spark documentation references the use of Pod Security Policies to restrict the privileges of Spark pods on Kubernetes. PodSecurityPolicy was [deprecated](https://kubernetes.io/blog/2021/04/08/kubernetes-1-21-release-announcement/#podsecuritypolicy-deprecation) in Kubernetes v1.21, and removed from Kubernetes in v1.25. The documentation should be updated to remove this reference and to direct users to the currently maintained Pod Security Admission Controller docs instead.

### Why are the changes needed?
Maintenance of the documentation to point to relevant and currently supported security features in Kubernetes.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Documentation changes only, no function changes to Spark.

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #53130 from Jimvin/master.

Authored-by: Jim Halfpenny <jim.halfpenny@stackable.tech>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>