[SPARK-54417][SQL] Fix error message for scalar subquery in IDENTIFIER clause
### What changes were proposed in this pull request?
Fix error message for scalar subquery in IDENTIFIER clause

### Why are the changes needed?
Previously, for a query like: `SELECT * FROM IDENTIFIER((SELECT 'tableName'))` we would throw an `INTERNAL_ERROR` because we never call `IdentifierResolution.evalIdentifierExpr` to check if the identifier expression is foldable. The call never happens, because in order to call `eval`, the expression has to be resolved first, but the subquery has `UnresolvedAlias('tableName')` that never gets resolved.

In `CheckAnalysis`, if the identifier expression is not resolved, I propose that we throw `NOT_A_CONSTANT_STRING.NOT_CONSTANT`, because ALL identifier expressions need to be constant and if by the end of analysis the expression is not resolved, that means it wasn't constant.

### Does this PR introduce _any_ user-facing change?
Error message improvement.

### How was this patch tested?
Added new test cases to golden file.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #53133 from mihailotim-db/mihailotim-db/fix_error_message_ident.

Authored-by: Mihailo Timotic <mihailo.timotic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>