[SPARK-51771][SQL][FOLLOWUP] Rename currentVersion to version in DSv2 Table
### What changes were proposed in this pull request?

This PR renames `currentVersion` to `version` in DSv2 `Table`. This method is supposed to be a simple getter and must not trigger a refresh of the underlying table state.

### Why are the changes needed?

These changes are needed to avoid ambiguity in the about to be released API in 4.1.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #53118 from aokolnychyi/spark-51771-followup.

Authored-by: Anton Okolnychyi <aokolnychyi@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>
(cherry picked from commit 6578b9b96e41bd15431d7ea7449683bb8d888bcd)
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>