[SPARK-54441][INFRA] Enable coverage data for workers
### What changes were proposed in this pull request?

By patching the worker in coverage runs, we save coverage data for code executed in workers. This code will only be triggered in coverage runs, not normal runs.

Also we uses the zip lib for many tests, this PR will also combine the coverage data with the `[paths]` config in `.coveragerc`.

### Why are the changes needed?

To have a better idea of what test coverage we are missing in our tests suite.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Manually test that coverage data shows up for the worker file.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #53145 from gaogaotiantian/enable-worker-coverage.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>