[SPARK-54364][INFRA] Separate test_install_spark from the rest of the tests
### What changes were proposed in this pull request?

We don't run `test_install_spark` in `pyspark_core` anymore - we consider it a separate module. We only run it when the related code changes, or once per day in `build_coverage`.

### Why are the changes needed?

It's a super slow test (2000~3000s) with almost no useful coverage. It basically downloads spark and unpack it. No one is touching the code and we spend hours every day to run this test.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

CI

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #53148 from gaogaotiantian/separate-install.

Authored-by: Tian Gao <gaogaotiantian@hotmail.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>