[SPARK-54448][SQL][DOCS] Fix docs and error message because CREATE TEMP VIEW does not support IF NOT EXISTS
### What changes were proposed in this pull request?

Currently, Spark SQL supports `CREATE [ OR REPLACE ] [ [ GLOBAL ] TEMPORARY ] VIEW [ IF NOT EXISTS ] view_identifier ...` syntax, but the implementation forbids using `CREATE [ GLOBAL ] TEMPORARY VIEW` with `IF NOT EXISTS`, this is likely an incompleted feature.

And this was also provided as an example in the docs, but actually does not work.

https://spark.apache.org/docs/4.0.1/sql-ref-syntax-ddl-create-view.html

<img width="890" height="366" alt="image" src="https://github.com/user-attachments/assets/d3a30055-866b-4307-904e-10e908b953a2" />

In addition, the error condition `TEMP_TABLE_OR_VIEW_ALREADY_EXISTS`'s message suggests "add the IF NOT EXISTS clause", which does not work either.

```
spark-sql (default)> CREATE GLOBAL TEMPORARY VIEW v AS SELECT 1;
spark-sql (default)> CREATE GLOBAL TEMPORARY VIEW v AS SELECT 1;
[TEMP_TABLE_OR_VIEW_ALREADY_EXISTS] Cannot create the temporary view `v` because it already exists.
Choose a different name, drop or replace the existing view,  or add the IF NOT EXISTS clause to tolerate pre-existing views. SQLSTATE: 42P07
```

### Why are the changes needed?

Fix wrong docs and error condition message.

### Does this PR introduce _any_ user-facing change?

It only updates the docs and error condition message, with no functionality changes.

### How was this patch tested?

<img width="928" height="1652" alt="image" src="https://github.com/user-attachments/assets/ec6d4a35-285a-4431-b456-60d2c4c8b1ac" />

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #53153 from pan3793/SPARK-54448.

Authored-by: Cheng Pan <chengpan@apache.org>
Signed-off-by: Dongjoon Hyun <dongjoon@apache.org>